Switch message size computation to u64
Attempt to serialize a message larger than 2G now explicitly results
in error.

+++ protobuf/src/doctest_pb.rs
@@ -71,10 +71,10 @@ impl crate::Message for MyMessage {
 
     // Compute sizes of nested messages
     #[allow(unused_variables)]
-    fn compute_size(&self) -> u32 {
+    fn compute_size(&self) -> u64 {
         let mut my_size = 0;
         my_size += crate::rt::unknown_fields_size(self.get_unknown_fields());
-        self.cached_size.set(my_size);
+        self.cached_size.set(my_size as u32);
         my_size
     }
